pyproject.toml
[tool.poetry]
name = "unichoose"
version = "0.1.0"
description = ""
authors = ["h4x4d <cleanyboom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=3.2.0,<3.3.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
haversine = "^2.7.0"
annoy = "^1.17.1"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^3.8.1"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"


[tool.poetry.group.dev.dependencies]
flake8 = ">=5.0.0,<6.0.0"
python-dotenv = "^0.21.0"
flake8-commas = "^2.1.0"
flake8-quotes = "^3.3.1"
flake8-comprehensions = "^3.10.1"
flake8-bugbear = "^22.12.6"
flake8-tidy-imports = "^4.8.0"
pep8-naming = "^0.13.2"
flake8-eradicate = "^1.4.0"
dlint = "^0.13.0"
flake8-print = "^5.0.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"
yapf = "^0.32.0"
toml = "^0.10.2"
requests = "^2.28.1"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
pyhamcrest = "^2.0.4"
allure-pytest = "^2.11.1"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
